Add in gantry and extruder measurements
These are my best-effort attempt to measure the bounding box for this odd, lumpy extruder assembly,
and there are serious concerns with the wiring harness given where it's attached and how it
protrudes from the assembly. However, I figured I should at least attempt to provide useful
guidelines for these dimensions to enable "One At A Time" printing if so-desired.